// x-pack/platform/plugins/shared/observability_ai_assistant/common/types.d.ts
export declare enum KnowledgeBaseEntryRole {
    AssistantSummarization = "assistant_summarization",
    UserEntry = "user_entry",
    Elastic = "elastic"
}
// ...
export interface KnowledgeBaseEntry {
    '@timestamp': string;
    id: string;
    title?: string;
    text: string;
    type?: 'user_instruction' | 'contextual';
    public: boolean;
    labels?: Record<string, string>;
    role: KnowledgeBaseEntryRole;
    user?: {
        name: string;
    };
    confidence?: 'low' | 'medium' | 'high';
    is_correction?: boolean;
}
export interface Instruction {
    id: string;
    text: string;
}